+++ src/components/icons/IconName.ts
@@ -1,9 +1,0 @@
-export type IconName =
-  | "cart"
-  | "external-link"
-  | "instagram"
-  | "facebook"
-  | "newsletter"
-  | "pinterest"
-  | "printer"
-  | "rss";