+++ src/icons/IconName.ts
@@ -1,0 +1,16 @@
+export const iconNames = [
+  "cart",
+  "close",
+  "copy",
+  "external-link",
+  "facebook",
+  "instagram",
+  "location",
+  "newsletter",
+  "pinterest",
+  "printer",
+  "restaurant",
+  "rss",
+];
+
+export type IconName = (typeof iconNames)[number];