+++ CHANGELOG.md
@@ -5,11 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.12]
+
+- added size attribute to button
+
+## [0.0.11]
+
+- added icons to storybook
+- unified handling of icons
+- updated khao-malet to v0.0.9
+
 ## [0.0.10]
 
 - added khao-social-buttons
 - added khao-icon-button component
-- udated khao-malet to v0.0.7
+- updated khao-malet to v0.0.7
 
 ## [0.0.0 - 0.0.9]